MAINT: rename test module
