+++ parser/src/ast/mod.rs
@@ -62,4 +62,4 @@ impl VarOrTerm {
 }
 
 #[cfg(test)]
-mod test;
+mod tests;
